- Updated the ImageProcessor sample to allow tiling or non-tiling sampler modes. - Added mouse wheel zoom to the ImageProcessor sample. - Updated the separable bilateral filter for the AmbientOcclusionI sample, so that large window sizes don't show artifacts on their edges. - Corrected an issue with the mouse wheel event processing, where all values came out positive before but are now corrected.

+++ trunk/Hieroglyph3/Source/Events/EvtMouseWheel.cpp
@@ -35,6 +35,6 @@ eEVENT EvtMouseWheel::GetEventType( )
 //--------------------------------------------------------------------------------
 int EvtMouseWheel::GetWheelDelta()
 {
-	return( ( 0xffff0000 & m_wparam ) >> 16 );
+	return( (short)(( 0xffff0000 & m_wparam ) >> 16 ) );
 }
 //--------------------------------------------------------------------------------
